refactor(types): split groups.ts into entities/group.ts
