+++ packages/types/src/__tests__/encryption.test.ts
@@ -70,6 +70,7 @@ import type { Channel } from "../entities/channel.js";
 import type { CustomFront } from "../entities/custom-front.js";
 import type { FrontingComment } from "../entities/fronting-comment.js";
 import type { FrontingSession } from "../entities/fronting-session.js";
+import type { Group } from "../entities/group.js";
 import type { JournalEntry } from "../entities/journal-entry.js";
 import type { LifecycleEvent } from "../entities/lifecycle-event.js";
 import type { MemberPhoto } from "../entities/member-photo.js";
@@ -81,7 +82,6 @@ import type { Poll } from "../entities/poll.js";
 import type { Relationship } from "../entities/relationship.js";
 import type { TimerConfig } from "../entities/timer-config.js";
 import type { WikiPage } from "../entities/wiki-page.js";
-import type { Group } from "../groups.js";
 import type {
   BucketId,
   FrontingCommentId,
